Fix local zkAS bootstrap and project detail route
